# openapi/iam2.v1.yaml
openapi: 3.0.3
info:
  title: IAM 2.0 API
  version: '1.0.0'
  description: >
    Versioned IAM 2.0 API consumed by NOX. All responses follow the normalized schema:
    {"ok": bool, "data": object, "errors": []}.
servers:
  - url: http://localhost:8010/v1
paths:
  /convert/molfile:
    post:
      summary: Convert MOL (V2000/V3000) to descriptors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConvertMolfileRequest'
      responses:
        '200':
          description: Normalized response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormalizedResponse'
...
components:
  schemas:
    NormalizedResponse:
      type: object
      required: [ok, data, errors]
      properties:
        ok:
          type: boolean
        data:
          type: object
          nullable: true
        errors:
          type: array
          items:
            type: string
    ConvertMolfileRequest:
      type: object
      required: [molfile]
      properties:
        molfile:
          type: string
          description: MOL V2000/V3000 content
...